refactor: move locale helpers to utils and dedupe Intl.Locale construction
`locale` was a folder for one file of stateless string helpers — same shape
as the rest of `shared/utils/`. Move it there.

While touching the file, extract a private `parseLocale` helper so
`normalizeLocale` and `getTextDirection` share one source of truth for the
POSIX-underscore repair, removing the double-parse in `getTextDirection`.
