+++ src/features/board/library/board-set-info-dialog.tsx
@@ -6,7 +6,7 @@ import DialogContent from "@mui/material/DialogContent";
 import DialogTitle from "@mui/material/DialogTitle";
 import Stack from "@mui/material/Stack";
 import Typography from "@mui/material/Typography";
-import { getEnglishLocaleName } from "@shared/locale/locale";
+import { getEnglishLocaleName } from "@shared/utils/locale";
 import type { BoardSetRecord } from "../storage/boards-db";
 
 export interface BoardSetInfoDialogProps {
